Remove `MarkTransformed`, `MarkSplit`, `MarkCausedByEvent`
Also type-specific command sequence factory methods to corresponding classes.
