+++ server/src/main/java/io/spine/server/command/model/CommandingMethod.java
@@ -37,10 +37,10 @@ import io.spine.type.MessageClass;
 import java.util.List;
 
 import static com.google.common.base.Preconditions.checkNotNull;
-import static io.spine.server.commandbus.CommandSequence.inResponseTo;
-import static io.spine.server.commandbus.CommandSequence.respondMany;
-import static io.spine.server.commandbus.CommandSequence.split;
-import static io.spine.server.commandbus.CommandSequence.transform;
+import static io.spine.server.commandbus.SeveralCommands.respondMany;
+import static io.spine.server.commandbus.SingleCommand.inResponseTo;
+import static io.spine.server.commandbus.Split.split;
+import static io.spine.server.commandbus.Transform.transform;
 
 /**
  * Base interface for methods that generate one or more command messages in response to
