+++ server/src/main/java/io/spine/server/commandbus/OnEvent.java
@@ -24,7 +24,7 @@ import com.google.protobuf.Message;
 import io.spine.core.ActorContext;
 import io.spine.core.CommandId;
 import io.spine.core.EventId;
-import io.spine.system.server.MarkCausedByEvent;
+import io.spine.system.server.CommandCausedByEvent;
 import io.spine.system.server.SystemGateway;
 
 /**
@@ -44,11 +44,11 @@ OnEvent <R extends Message, B extends Message.Builder, S extends CommandSequence
     }
 
     void markReacted(SystemGateway gateway, CommandId commandId) {
-        MarkCausedByEvent systemCommand = MarkCausedByEvent
+        CommandCausedByEvent systemEvent = CommandCausedByEvent
                 .newBuilder()
                 .setId(commandId)
                 .setOrigin(origin())
                 .build();
-        gateway.postCommand(systemCommand);
+        gateway.postEvent(systemEvent);
     }
 }
