+++ server/src/main/proto/spine/system/server/c/command_lifecycle_commands.proto
@@ -63,47 +63,3 @@ message AssignTargetToCommand {
     // The command target.
     CommandTarget target = 2 [(valid) = true];
 }
-
-// Marks a command as one which was transformed into another command.
-//
-// Posted when a commanding method created one command in response to an incoming one.
-// For the case of more than one produced command, please use `MarkSplit`.
-//
-message MarkTransformed {
-
-    // The ID of the command.
-    spine.core.CommandId id = 1;
-
-    // The ID of the command generated in response to the received command.
-    spine.core.CommandId produced = 2 [(required) = true];
-}
-
-// Marks a command which cased two or more commands.
-//
-// Posted when a commanding method created two or more commands in response to an incoming one.
-//
-message MarkSplit {
-
-    // The ID of the command.
-    spine.core.CommandId id = 1;
-
-    // Identifiers of commands generated in response to the `source` command.
-    //
-    // There must be two or more produced commands.
-    // For one produced command please use `MarkTransformed`.
-    //
-    repeated core.CommandId produced = 2 [(required) = true];
-}
-
-// Mark a command as one which was created in response to an event.
-//
-// Posted when a commanding method creates one or more commands in response to an incoming event.
-//
-message MarkCausedByEvent {
-
-    // The ID of the command.
-    spine.core.CommandId id = 1;
-
-    // The ID of the event which cased the command.
-    spine.core.EventId origin = 2;
-}
